+++ pkg/storage/storage.go
@@ -2,7 +2,7 @@ package storage
 
 // Post - публикация.
 type Post struct {
-	ID       int64
+	ID       int
 	Title    string
 	Content  string
 	PostTime int64
@@ -12,5 +12,5 @@ type Post struct {
 // Interface задаёт контракт на работу с БД.
 type Interface interface {
 	GetPosts(n int) ([]Post, error) // Возвращает список из n публикаций
-	AddPost(Post) (int64, error)    // Добавляет новую публикацию
+	AddPost(Post) (int, error)      // Добавляет новую публикацию
 }
